% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_regimes.R
\name{make_regime_prescribed}
\alias{make_regime_prescribed}
\title{Define a prescribed fire regime with random cell selection}
\usage{
make_regime_prescribed(name, fn_prob_tsf, fn_occur, prop_landscape)
}
\description{
Define a prescribed fire regime with random cell selection
}
