% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_regimes.R
\name{make_regime_wildfire}
\alias{make_regime_wildfire}
\title{Define a wildfire regime}
\usage{
make_regime_wildfire(
  name,
  fn_prob_tsf,
  fn_occur,
  diagonal = FALSE,
  max_ignition_attempts = 1
)
}
\description{
Define a wildfire regime
}
