% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_sim.R
\name{rcafe_simulate}
\alias{rcafe_simulate}
\title{Run an RCAFE simulation}
\usage{
rcafe_simulate(
  DB_path,
  tsf_init,
  regimes,
  n_rep = 1,
  n_iter = 100,
  n_burnin = 0,
  tsf_breaks = c(5, 10, 20),
  tsf_map_times = NULL,
  map_fire_types = c("none", "all", "wild", "prescribed"),
  display_progress = TRUE
)
}
\arguments{
\item{DB_path}{Path to a folder for the DuckDB database file to which
simulation results will be recorded.}

\item{tsf_init}{A matrix of of initial time since fire values for the
landscape. The dimensions of this matrix will be used to define those of
the simulation landscape. Values should be non-negative integers.}

\item{regimes}{A list of one or definitions for wildfire and/or prescribed
fire regime(s).}

\item{n_rep}{(positive integer; default 1) Number of replicate simulations.}

\item{n_iter}{(positive integer; default 100) Number of iterations (time
steps) per simulation.}

\item{n_burnin}{(non-negative integer; default 0) Number of initial time
steps to run the simulation for prior to the \code{n_iter} times. No data
are recorded for these initial time steps. This can be useful to evolve an
initial, spatially-structured pattern of time since fire (see examples).}

\item{tsf_breaks}{An integer vector specifying the breakpoints over which to
record the proportions of the simulation landscape within time since fire
intervals at each time step (year). For example, the default value of
\code{tsf_breaks = c(5, 10, 20)} specifies to record landscape proportions
in the intervals 0-5 years;  6-10 years; 10-20 years; and greater than 20
years.}

\item{tsf_map_times}{\strong{Not implemented yet}. An integer vector
specifying the times at which to record a map of time since fire. Maps will
only be recorded after any burn-in time steps (argument \code{n_burnin})
have been completed. For example, the arguments
\code{n_iter=500, n_burnin=100, tsf_map_times = c(1, seq(100, 500,
50))} specify to run 100 burn-in simulations to evolve a time since fire
pattern in the landscape, and then record a map in the first year and every
hundred years of the simulation period. If this vector is missing,
\code{NULL} or empty, no time since fire maps will be saved.}

\item{map_fire_types}{(character; default \code{'none'}) Whether to save maps
of fires in the output database. Can be one of the following: \code{'all'}
(save both prescribed and wild fires); \code{'wild'} (save only wildfires);
\code{'prescribed'} (save only prescribed fires); \code{'none'} (do not
save fire maps). Can be abbreviated (e.g. \code{map_fire_types = 'w'}).}

\item{display_progress}{(logical) Whether to display a progress bar.}
}
\description{
Run an RCAFE simulation
}
\examples{
\dontrun{
# Function to relate cell burning probability to time since fire
fn_prob <- function(tsf) {
  1 / (1 + exp(1.65 - 0.16*tsf))
}

# Visualize the function for a range of time since fire values
curve(fn_prob(x), from=0, to=40,
      ylim = c(0, 1), xlab = "TSF", ylab = "Burn probability")

# Define a 100 x 100 landscape with uniform initial time since fire
# of 20 years
tsf0 <- matrix(20, nrow=100, ncol=100)

# Run the simulation for 500 years after an initial 100 year
# burn-in period
res <- rcafe_simulate(tsf0, n_iter=500, n_burnin=100,
                      fn_prob_tsf = fn_prob)
}

}
